backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
python-multipart==0.0.9

# Async Tasks
celery==5.4.0
redis==5.0.4

# File Storage
minio==7.2.7

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# SMS
twilio==9.1.0

# PDF Generation (for UC & Invoices)
reportlab==4.2.0
weasyprint==62.1

# Utilities
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0

# Geo & Maps
geopy==2.4.1
shapely==2.0.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
qrcode[pil]==7.4.2

qrcode[pil]==7.4.2
